Start PhotonVision after network.target is reached (#114)
PV tries to access networking information fairly early in the startup
process and, when NetworkManager isn't ready, PV cannot get the
information it needs about the network hardware.

This PR configures the photonvision service to start after
NetworkManager is running. The change is only made when PV is configured
to manage networking.

This should reduce the need for
https://github.com/PhotonVision/photonvision/pull/2240, but it's safer
to have both.